committing pipeline code for reference
contains the new pipeline code for extracting trial codes from the beh data

there is a better way of doing this while retaining backwards compatibility by using allCodes instead of e.id but this is cleaner -- if not relying on codes, don't rely on codes.
major caveat: this method assumes that the first start code is legit because it references trial start times from beh to that code. if the data looks like junk for any experiment, this is likely the issue.

also contains neuroshare dependency (this is the old way of reading nev etc.)